🧑‍💻 refactor infinteTranslateX into hook
